moved dataset reader
