+++ allennlp/data/dataset_readers/__init__.py
@@ -31,8 +31,7 @@ from allennlp.data.dataset_readers.stanford_sentiment_tree_bank import (
         StanfordSentimentTreeBankDatasetReader)
 from allennlp.data.dataset_readers.quora_paraphrase import QuoraParaphraseDatasetReader
 from allennlp.data.dataset_readers.semantic_parsing import (
-        WikiTablesDatasetReader, WikiTablesVariableFreeDatasetReader, AtisDatasetReader, NlvrDatasetReader,
-        TemplateText2SqlDatasetReader)
+        WikiTablesDatasetReader, AtisDatasetReader, NlvrDatasetReader, TemplateText2SqlDatasetReader)
 from allennlp.data.dataset_readers.semantic_parsing.quarel import QuarelDatasetReader
 from allennlp.data.dataset_readers.simple_language_modeling import SimpleLanguageModelingDatasetReader
 from allennlp.data.dataset_readers.babi import BabiReader
