+++ allennlp/data/dataset_readers/semantic_parsing/__init__.py
@@ -2,6 +2,5 @@
 from allennlp.data.dataset_readers.semantic_parsing.atis import AtisDatasetReader
 from allennlp.data.dataset_readers.semantic_parsing.nlvr import NlvrDatasetReader
 from allennlp.data.dataset_readers.semantic_parsing.wikitables.wikitables import WikiTablesDatasetReader
-from allennlp.data.dataset_readers.semantic_parsing.wikitables.wikitables_variable_free import WikiTablesVariableFreeDatasetReader
 from allennlp.data.dataset_readers.semantic_parsing.template_text2sql import TemplateText2SqlDatasetReader
 from allennlp.data.dataset_readers.semantic_parsing.grammar_based_text2sql import GrammarBasedText2SqlDatasetReader
